+++ java-racingcar-playground/src/main/java/racing_car/domain/Car.java
@@ -1,4 +1,4 @@
-package racing_car;
+package racing_car.domain;
 
 public class Car {
