+++ java-racingcar-playground/src/main/java/racing_car/domain/CarName.java
@@ -1,4 +1,4 @@
-package racing_car;
+package racing_car.domain;
 
 public class CarName {
