+++ java-racingcar-playground/src/test/java/racing_car/domain/MovingTest.java
@@ -1,10 +1,11 @@
-package racing_car;
+package racing_car.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing_car.domain.Moving;
 
 public class MovingTest {
